+++ EpicLoot/Data/EIDFLegacy.cs
@@ -1,4 +1,7 @@
 ﻿
+using System;
+using System.Linq;
+
 namespace EpicLoot.Data
 {
     public static class EIDFLegacy
@@ -20,6 +23,27 @@ namespace EpicLoot.Data
         public static bool IsLegacyMagicItem(this ItemDrop.ItemData itemData)
         {
             return itemData.m_crafterName.Contains($"{StartDelimiter}{MagicItemComponent.TypeID}{EndDelimiter}");
+        }
+
+        public static string GetMagicItemFromCrafterName(ItemDrop.ItemData item)
+        {
+            var serializedComponents = item.m_crafterName.Split(new[] { StartDelimiter }, StringSplitOptions.RemoveEmptyEntries);
+
+            foreach (var component in serializedComponents)
+            {
+                var parts = component.Split(new[] { EndDelimiter }, StringSplitOptions.None);
+                var typeString = RestoreDataText(parts[0]);
+
+                if (typeString.Equals(MagicItemComponent.TypeID))
+                {
+                    var data = parts.Length == 2 ? parts[1] : string.Empty;
+                    if (string.IsNullOrEmpty(data))
+                        continue;
+
+                    return RestoreDataText(data);
+                }
+            }
+            return null;
         }
     }
 }
